重建为更简单的syscall

+++ include/string.h
@@ -20,7 +20,7 @@
 #endif
 
 /* Standard C library string functions */
-int strlen(const char *s);
+size_t strlen(const char *s);
 int strcmp(const char *s1, const char *s2);
 int strncmp(const char *s1, const char *s2, size_t n);
 char *strcpy(char *dest, const char *src);
